fix:delete embedders utils
