+++ src/rs_embed/embedders/meta.py
