+++ src/rs_embed/embedders/onthefly_agrifm.py
@@ -27,16 +27,20 @@ from ..core.specs import (
     TemporalSpec,
 )
 from ..providers import ProviderBase
-from ._vit_mae_utils import ensure_torch
+def ensure_torch() -> None:
+    try:
+        import torch  # noqa: F401
+    except Exception as e:
+        raise ModelError("This embedder requires torch installed.") from e
 from .base import EmbedderBase
-from .meta_utils import build_meta, temporal_to_range
-from .runtime_utils import (
+from .meta import build_meta, temporal_to_range
+from ..providers.fetch import (
     fetch_s2_multiframe_raw_tchw as _fetch_s2_multiframe_raw_tchw,
 )
-from .runtime_utils import (
+from ..providers.resolution import (
     is_provider_backend,
 )
-from .runtime_utils import (
+from ..tools.runtime import (
     load_cached_with_device as _load_cached_with_device,
 )
