+++ src/rs_embed/embedders/onthefly_galileo.py
@@ -24,22 +24,24 @@ from ..core.specs import (
 )
 from ..providers import ProviderBase
 from ..tools.temporal import temporal_frame_midpoints
-from ._vit_mae_utils import ensure_torch
+def ensure_torch() -> None:
+    try:
+        import torch  # noqa: F401
+    except Exception as e:
+        raise ModelError("This embedder requires torch installed.") from e
 from .base import EmbedderBase
-from .meta_utils import build_meta, temporal_to_range
-from .runtime_utils import (
-    coerce_input_to_tchw as _coerce_input_to_tchw,
-)
-from .runtime_utils import (
+from .meta import build_meta, temporal_to_range
+from ..providers.fetch import (
     fetch_s2_multiframe_raw_tchw as _fetch_s2_multiframe_raw_tchw,
 )
-from .runtime_utils import (
+from ..providers.resolution import (
     is_provider_backend,
 )
-from .runtime_utils import (
-    load_cached_with_device as _load_cached_with_device,
+from ..tools.normalization import (
+    coerce_input_to_tchw as _coerce_input_to_tchw,
 )
-from .runtime_utils import (
+from ..tools.runtime import (
+    load_cached_with_device as _load_cached_with_device,
     resolve_device_auto_torch as _resolve_device,
 )
