+++ src/rs_embed/embedders/onthefly_prithvi.py
@@ -21,23 +21,63 @@ from ..core.specs import (
     TemporalSpec,
 )
 from ..providers import ProviderBase
-from ._vit_mae_utils import (
-    base_meta,
-    ensure_torch,
-    pool_from_tokens,
-    temporal_to_range,
-    tokens_to_grid_dhw,
-)
+
+def ensure_torch() -> None:
+    try:
+        import torch  # noqa: F401
+    except Exception as e:
+        raise ModelError("This embedder requires torch installed.") from e
+
+from .meta import build_meta, temporal_to_range
+
+def base_meta(*, model_name, hf_id, backend, image_size, sensor,
+              temporal=None, source=None, embed_type="on_the_fly", extra=None):
+    m = build_meta(
+        model=model_name, kind=embed_type, backend=backend,
+        source=source or getattr(sensor, "collection", None),
+        sensor=sensor, temporal=temporal, image_size=image_size,
+    )
+    m["hf_id"] = hf_id
+    if extra:
+        m.update(extra)
+    return m
+
+
+def pool_from_tokens(tokens, pooling):
+    n = len(tokens)
+    h2 = int((n - 1) ** 0.5)
+    has_cls = n > 1 and h2 * h2 == n - 1
+    patch = tokens[1:] if has_cls else tokens
+    if len(patch) == 0:
+        return tokens[0].astype("float32"), has_cls
+    if pooling == "mean":
+        return patch.mean(axis=0).astype("float32"), has_cls
+    if pooling == "max":
+        return patch.max(axis=0).astype("float32"), has_cls
+    raise ModelError(f"Unknown pooling={pooling!r} (expected 'mean' or 'max').")
+
+
+def tokens_to_grid_dhw(tokens):
+    n = len(tokens)
+    h2 = int((n - 1) ** 0.5)
+    has_cls = n > 1 and h2 * h2 == n - 1
+    patch = tokens[1:] if has_cls else tokens
+    p, d = patch.shape
+    hw = int(p ** 0.5)
+    if hw * hw != p:
+        raise ModelError(f"Patch token count {p} is not a perfect square.")
+    return patch.reshape(hw, hw, d).transpose(2, 0, 1).astype("float32"), (hw, hw), has_cls
+
 from .base import EmbedderBase
-from .config_utils import model_config_value
-from .meta_utils import temporal_midpoint_str
-from .runtime_utils import (
+from .config import model_config_value
+from .meta import temporal_midpoint_str
+from ..providers.fetch import (
     fetch_collection_patch_chw as _fetch_collection_patch_chw,
 )
-from .runtime_utils import (
+from ..providers.resolution import (
     is_provider_backend,
 )
-from .runtime_utils import (
+from ..tools.runtime import (
     load_cached_with_device as _load_cached_with_device,
 )
 
@@ -509,7 +549,7 @@ class PrithviEOV2S2_6B_Embedder(EmbedderBase):
       - temporal: range/year (year->full year)
       - sensor: controls provider fetch (scale/cloudy/composite)
 
-    Outputs (aligned with _vit_mae_utils):
+    Outputs:
       - pooled: patch-token mean/max (exclude CLS if present)
       - grid: token map [D,H,W] (exclude CLS if present)
     """
