+++ src/rs_embed/embedders/onthefly_satmaepp_s2.py
@@ -20,16 +20,36 @@ from ..core.specs import (
     TemporalSpec,
 )
 from ..providers.base import ProviderBase
-from ._vit_mae_utils import base_meta, ensure_torch, temporal_to_range
+
+def ensure_torch() -> None:
+    try:
+        import torch  # noqa: F401
+    except Exception as e:
+        raise ModelError("This embedder requires torch installed.") from e
+
+from .meta import build_meta, temporal_to_range
+
+def base_meta(*, model_name, hf_id, backend, image_size, sensor,
+              temporal=None, source=None, embed_type="on_the_fly", extra=None):
+    m = build_meta(
+        model=model_name, kind=embed_type, backend=backend,
+        source=source or getattr(sensor, "collection", None),
+        sensor=sensor, temporal=temporal, image_size=image_size,
+    )
+    m["hf_id"] = hf_id
+    if extra:
+        m.update(extra)
+    return m
+
 from .base import EmbedderBase
-from .config_utils import model_config_value
-from .runtime_utils import (
+from .config import model_config_value
+from ..providers.fetch import (
     fetch_collection_patch_chw as _fetch_collection_patch_chw,
 )
-from .runtime_utils import (
+from ..providers.resolution import (
     is_provider_backend,
 )
-from .runtime_utils import (
+from ..tools.runtime import (
     load_cached_with_device as _load_cached_with_device,
 )
