+++ src/rs_embed/embedders/onthefly_terramind.py
@@ -21,19 +21,51 @@ from ..core.specs import (
     TemporalSpec,
 )
 from ..providers import ProviderBase
-from ._vit_mae_utils import ensure_torch, pool_from_tokens, tokens_to_grid_dhw
+
+def ensure_torch() -> None:
+    try:
+        import torch  # noqa: F401
+    except Exception as e:
+        raise ModelError("This embedder requires torch installed.") from e
+
+
+def pool_from_tokens(tokens, pooling):
+    n = len(tokens)
+    h = int(n ** 0.5 - 1) if n > 1 else 0
+    # check h+1 squared equals n for CLS detection
+    h2 = int((n - 1) ** 0.5)
+    has_cls = n > 1 and h2 * h2 == n - 1
+    patch = tokens[1:] if has_cls else tokens
+    if len(patch) == 0:
+        return tokens[0].astype("float32"), has_cls
+    if pooling == "mean":
+        return patch.mean(axis=0).astype("float32"), has_cls
+    if pooling == "max":
+        return patch.max(axis=0).astype("float32"), has_cls
+    raise ModelError(f"Unknown pooling={pooling!r} (expected 'mean' or 'max').")
+
+
+def tokens_to_grid_dhw(tokens):
+    n = len(tokens)
+    h2 = int((n - 1) ** 0.5)
+    has_cls = n > 1 and h2 * h2 == n - 1
+    patch = tokens[1:] if has_cls else tokens
+    p, d = patch.shape
+    hw = int(p ** 0.5)
+    if hw * hw != p:
+        raise ModelError(f"Patch token count {p} is not a perfect square.")
+    return patch.reshape(hw, hw, d).transpose(2, 0, 1).astype("float32"), (hw, hw), has_cls
+
 from .base import EmbedderBase
-from .meta_utils import build_meta, temporal_to_range
-from .runtime_utils import (
-    coerce_single_input_chw,
-)
-from .runtime_utils import (
+from .meta import build_meta, temporal_to_range
+from ..providers.fetch import (
     fetch_collection_patch_chw as _fetch_collection_patch_chw,
 )
-from .runtime_utils import (
-    load_cached_with_device as _load_cached_with_device,
+from ..tools.normalization import (
+    coerce_single_input_chw,
 )
-from .runtime_utils import (
+from ..tools.runtime import (
+    load_cached_with_device as _load_cached_with_device,
     resolve_device_auto_torch as _resolve_device,
 )
