+++ src/rs_embed/providers/fetch.py
@@ -1,145 +1,37 @@
-"""Provider runtime helpers for backend choice, provider fetch, and device detection."""
+"""Provider fetch helpers and satellite-data normalization."""
 
 from __future__ import annotations
 
-import os
 from collections.abc import Callable, Sequence
-from typing import Any, TypeVar
+from typing import Any
 
 import numpy as np
 
 from ..core.errors import ModelError
 from ..core.specs import NormalizationSpec, SensorSpec, SpatialSpec, TemporalSpec
-from ..providers import get_provider, has_provider, list_providers
-from ..providers.base import ProviderBase
-from ..tools.normalization import normalize_backend_name
-
-_T = TypeVar("_T")
+from .base import ProviderBase
 
 
-def default_provider_backend_name() -> str | None:
-    configured = normalize_backend_name(os.environ.get("RS_EMBED_DEFAULT_PROVIDER", ""))
-    if configured:
-        return configured if has_provider(configured) else None
-    providers = list_providers()
-    if not providers:
-        return None
-    if "gee" in providers:
-        return "gee"
-    return str(providers[0]).strip().lower()
-
-
-def resolve_provider_backend_name(
-    backend: str,
+def fetch_sensor_patch_chw(
+    provider: ProviderBase,
     *,
-    allow_auto: bool = True,
-    auto_backend: str | None = None,
-) -> str | None:
-    b = normalize_backend_name(backend)
-    if allow_auto and b == "auto":
-        resolved_auto = (
-            normalize_backend_name(auto_backend)
-            if auto_backend is not None
-            else default_provider_backend_name()
-        )
-        if not resolved_auto:
-            return None
-        b = resolved_auto
-    if has_provider(b):
-        return b
-    return None
+    spatial: SpatialSpec,
+    temporal: TemporalSpec | None,
+    sensor: SensorSpec,
+    to_float_image: bool = False,
+) -> np.ndarray:
+    """Fetch a CHW patch from a concrete SensorSpec, re-raising ProviderError as ModelError."""
+    from ..core.errors import ProviderError
 
-
-def is_provider_backend(
-    backend: str,
-    *,
-    allow_auto: bool = True,
-    auto_backend: str | None = None,
-) -> bool:
-    return (
-        resolve_provider_backend_name(
-            backend,
-            allow_auto=allow_auto,
-            auto_backend=auto_backend,
-        )
-        is not None
-    )
-
-
-def get_cached_provider(
-    provider_cache: dict[str, ProviderBase],
-    *,
-    backend: str,
-    allow_auto: bool = True,
-    auto_backend: str | None = None,
-) -> ProviderBase:
-    b = resolve_provider_backend_name(
-        backend,
-        allow_auto=allow_auto,
-        auto_backend=auto_backend,
-    )
-    if b is None:
-        raise ModelError(f"Unsupported provider backend={backend!r}.")
-    p = provider_cache.get(b)
-    if p is None:
-        kwargs = provider_init_kwargs(b)
-        p = get_provider(b, **kwargs)
-        provider_cache[b] = p
-    p.ensure_ready()
-    return p
-
-
-def provider_init_kwargs(backend: str) -> dict[str, Any]:
-    """Provider-specific constructor kwargs, centralized outside embedders."""
-    b = normalize_backend_name(backend)
-    if b == "gee":
-        return {"auto_auth": True}
-    return {}
-
-
-def create_provider_for_backend(
-    backend: str,
-    *,
-    allow_auto: bool = True,
-    auto_backend: str | None = None,
-) -> ProviderBase:
-    b = resolve_provider_backend_name(
-        backend,
-        allow_auto=allow_auto,
-        auto_backend=auto_backend,
-    )
-    if b is None:
-        raise ModelError(f"Unsupported provider backend={backend!r}.")
-    p = get_provider(b, **provider_init_kwargs(b))
-    p.ensure_ready()
-    return p
-
-
-def resolve_device_auto_torch(device: str) -> str:
-    if device != "auto":
-        return device
     try:
-        import torch
-
-        if torch.cuda.is_available():
-            return "cuda"
-        if torch.backends.mps.is_available():
-            return "mps"
-        return "cpu"
-    except Exception as _e:
-        return "cpu"
-
-
-def load_cached_with_device(
-    cached_loader: Callable[..., _T],
-    *,
-    device: str,
-    **kwargs: Any,
-) -> tuple[_T, str]:
-    """Resolve device once and call a cached loader that accepts `dev=...`."""
-    dev = resolve_device_auto_torch(device)
-    loaded = cached_loader(dev=dev, **kwargs)
-    return loaded, dev
+        return provider.fetch_sensor_patch_chw(
+            spatial=spatial,
+            temporal=temporal,
+            sensor=sensor,
+            to_float_image=to_float_image,
+        )
+    except ProviderError as exc:
+        raise ModelError(str(exc)) from exc
 
 
 def fetch_collection_patch_chw(
@@ -171,33 +63,6 @@ def fetch_collection_patch_chw(
     )
 
 
-def fetch_sensor_patch_chw(
-    provider: ProviderBase,
-    *,
-    spatial: SpatialSpec,
-    temporal: TemporalSpec | None,
-    sensor: SensorSpec,
-    to_float_image: bool = False,
-) -> np.ndarray:
-    """Fetch a CHW patch from a concrete SensorSpec.
-
-    Delegates to ``provider.fetch_sensor_patch_chw`` (which already validates
-    ndim, channel count, and calls nan_to_num).  Errors are re-raised as
-    ModelError so embedder-layer callers see a consistent exception type.
-    """
-    from ..core.errors import ProviderError
-
-    try:
-        return provider.fetch_sensor_patch_chw(
-            spatial=spatial,
-            temporal=temporal,
-            sensor=sensor,
-            to_float_image=to_float_image,
-        )
-    except ProviderError as exc:
-        raise ModelError(str(exc)) from exc
-
-
 def _stitch_spatial_last2_arrays(
     *,
     a: np.ndarray,
@@ -207,7 +72,7 @@ def _stitch_spatial_last2_arrays(
     scale_m: int,
     fill_value: float,
 ) -> np.ndarray:
-    from ..providers.gee_utils import _stitch_bbox_split_arrays
+    from .gee_utils import _stitch_bbox_split_arrays
 
     return _stitch_bbox_split_arrays(
         arr_a=np.asarray(a, dtype=np.float32),
@@ -228,7 +93,7 @@ def _fetch_spatial_array_with_bbox_fallback(
     fetch_fn: Callable[[SpatialSpec], np.ndarray],
     split_depth: int = 0,
 ) -> np.ndarray:
-    from ..providers import gee_utils as _ah
+    from . import gee_utils as _ah
 
     try:
         return np.asarray(fetch_fn(spatial), dtype=np.float32)
@@ -302,7 +167,7 @@ def fetch_collection_patch_all_bands_chw(
         arr, names = _fetch_once(spatial)
         return np.asarray(arr, dtype=np.float32), tuple(names)
     except Exception as e:
-        from ..providers import gee_utils as _ah
+        from . import gee_utils as _ah
 
         if not (
             _ah._looks_like_gee_sample_too_many_pixels(e) and _ah._looks_like_bbox_spatial(spatial)
@@ -461,46 +326,6 @@ def fetch_s1_vvvh_raw_chw_with_meta(
     return np.nan_to_num(arr, nan=0.0, posinf=0.0, neginf=0.0).astype(np.float32), meta_out
 
 
-def apply_normalization(raw: np.ndarray, norm: NormalizationSpec) -> np.ndarray:
-    """Apply a declarative normalization strategy to raw provider data.
-
-    Parameters
-    ----------
-    raw : np.ndarray
-        Raw provider array (CHW or TCHW).
-    norm : NormalizationSpec
-        Normalization strategy to apply.
-
-    Returns
-    -------
-    np.ndarray
-        Normalized float32 array.
-    """
-    arr = np.asarray(raw, dtype=np.float32)
-    if norm.mode == "s2_sr_clip":
-        return np.clip(arr / 10000.0, 0.0, 1.0).astype(np.float32)
-    if norm.mode == "s2_sr_raw":
-        return np.clip(arr, 0.0, 10000.0).astype(np.float32)
-    if norm.mode == "s1_log_normalize":
-        return normalize_s1_vvvh_chw(arr)
-    if norm.mode == "none":
-        return np.nan_to_num(arr, nan=0.0, posinf=0.0, neginf=0.0).astype(np.float32)
-    raise ModelError(f"Unknown normalization mode: {norm.mode!r}")
-
-
-def normalize_s1_vvvh_chw(raw_chw: np.ndarray) -> np.ndarray:
-    """Convert raw S1 VV/VH to numerically stable [0,1] CHW."""
-    arr = np.asarray(raw_chw, dtype=np.float32)
-    if arr.ndim != 3 or int(arr.shape[0]) != 2:
-        raise ModelError(
-            f"Expected raw S1 VV/VH CHW with C=2, got shape={getattr(arr, 'shape', None)}"
-        )
-    x = np.log1p(np.maximum(arr, 0.0))
-    denom = np.percentile(x, 99) if np.isfinite(x).all() else 1.0
-    denom = float(denom) if float(denom) > 0 else 1.0
-    return np.clip(x / denom, 0.0, 1.0).astype(np.float32)
-
-
 def fetch_s2_multiframe_raw_tchw(
     provider: ProviderBase,
     *,
@@ -542,78 +367,28 @@ def fetch_s2_multiframe_raw_tchw(
     return np.nan_to_num(arr, nan=0.0, posinf=0.0, neginf=0.0).astype(np.float32)
 
 
-def coerce_input_to_tchw(
-    input_chw: np.ndarray,
-    *,
-    expected_channels: int,
-    n_frames: int,
-    model_name: str,
-) -> np.ndarray:
-    """Normalize user-provided CHW/TCHW into clipped float32 [T,C,H,W]."""
-    raw = np.asarray(input_chw, dtype=np.float32)
-    t = max(1, int(n_frames))
-
-    if raw.ndim == 3:
-        if int(raw.shape[0]) != int(expected_channels):
-            raise ModelError(
-                f"input_chw must be CHW with C={int(expected_channels)} for {model_name}, "
-                f"got {tuple(int(v) for v in raw.shape)}"
-            )
-        raw_tchw = np.repeat(raw[None, ...], repeats=t, axis=0).astype(np.float32)
-    elif raw.ndim == 4:
-        if int(raw.shape[1]) != int(expected_channels):
-            raise ModelError(
-                f"input_chw must be TCHW with C={int(expected_channels)} for {model_name}, "
-                f"got {tuple(int(v) for v in raw.shape)}"
-            )
-        raw_tchw = raw.astype(np.float32, copy=False)
-        if raw_tchw.shape[0] < t:
-            raw_tchw = np.concatenate(
-                [raw_tchw] + [raw_tchw[-1:]] * (t - raw_tchw.shape[0]),
-                axis=0,
-            )
-        elif raw_tchw.shape[0] > t:
-            raw_tchw = raw_tchw[:t]
-    else:
+def normalize_s1_vvvh_chw(raw_chw: np.ndarray) -> np.ndarray:
+    """Convert raw S1 VV/VH to numerically stable [0,1] CHW."""
+    arr = np.asarray(raw_chw, dtype=np.float32)
+    if arr.ndim != 3 or int(arr.shape[0]) != 2:
         raise ModelError(
-            f"input_chw must be CHW (C,H,W) or TCHW (T,C,H,W) for {model_name}, "
-            f"got {tuple(int(v) for v in raw.shape)}"
+            f"Expected raw S1 VV/VH CHW with C=2, got shape={getattr(arr, 'shape', None)}"
         )
+    x = np.log1p(np.maximum(arr, 0.0))
+    denom = np.percentile(x, 99) if np.isfinite(x).all() else 1.0
+    denom = float(denom) if float(denom) > 0 else 1.0
+    return np.clip(x / denom, 0.0, 1.0).astype(np.float32)
 
-    raw_tchw = np.nan_to_num(raw_tchw, nan=0.0, posinf=0.0, neginf=0.0)
-    return np.clip(raw_tchw, 0.0, 10000.0).astype(np.float32)
 
-
-def coerce_single_input_chw(
-    input_chw: Any,
-    *,
-    expected_channels: int | None,
-    model_name: str,
-) -> np.ndarray:
-    """Normalize one user-provided tensor input into float32 CHW."""
-    raw = input_chw
-    try:
-        import torch
-
-        if torch.is_tensor(raw):
-            raw = raw.detach().cpu().numpy()
-    except Exception as _e:
-        pass
-
+def apply_normalization(raw: np.ndarray, norm: NormalizationSpec) -> np.ndarray:
+    """Apply a declarative normalization strategy to raw provider data."""
     arr = np.asarray(raw, dtype=np.float32)
-    if arr.ndim == 4:
-        raise ModelError(
-            f"{model_name} expects single-sample input_chw as CHW (C,H,W), "
-            f"got {tuple(int(v) for v in arr.shape)}. "
-            "Use get_embeddings_batch_from_inputs(...) for batches."
-        )
-    if arr.ndim != 3:
-        raise ModelError(
-            f"{model_name} expects input_chw as CHW (C,H,W), got {tuple(int(v) for v in arr.shape)}"
-        )
-    if expected_channels is not None and int(arr.shape[0]) != int(expected_channels):
-        raise ModelError(
-            f"input_chw must be CHW with C={int(expected_channels)} for {model_name}, "
-            f"got {tuple(int(v) for v in arr.shape)}"
-        )
-    return np.nan_to_num(arr, nan=0.0, posinf=0.0, neginf=0.0).astype(np.float32)
+    if norm.mode == "s2_sr_clip":
+        return np.clip(arr / 10000.0, 0.0, 1.0).astype(np.float32)
+    if norm.mode == "s2_sr_raw":
+        return np.clip(arr, 0.0, 10000.0).astype(np.float32)
+    if norm.mode == "s1_log_normalize":
+        return normalize_s1_vvvh_chw(arr)
+    if norm.mode == "none":
+        return np.nan_to_num(arr, nan=0.0, posinf=0.0, neginf=0.0).astype(np.float32)
+    raise ModelError(f"Unknown normalization mode: {norm.mode!r}")
