+++ tests/test_meta_utils.py
@@ -1,5 +1,5 @@
 from rs_embed.core.specs import TemporalSpec
-from rs_embed.embedders.meta_utils import temporal_to_dict, temporal_to_range
+from rs_embed.embedders.meta import temporal_to_dict, temporal_to_range
 from rs_embed.tools.temporal import split_temporal_range, temporal_frame_midpoints
